Add compatibility wrapper for phase4_functions
