+++ cluster_confusion_heatmap.py
@@ -8,7 +8,7 @@ from pathlib import Path
 
 import pandas as pd
 
-from phase4_functions import (
+from phase4.functions import (
     cluster_confusion_table,
     plot_cluster_confusion_heatmap,
 )
