+++ compare_pca_versions.py
@@ -14,7 +14,7 @@ import pandas as pd
 import seaborn as sns
 import yaml
 
-from phase4_functions import (
+from phase4.functions import (
     load_datasets,
     prepare_data,
     select_variables,
