+++ export_mfa_inertias.py
@@ -10,7 +10,7 @@ import pandas as pd
 import yaml
 
 import phase4
-from phase4_functions import (
+from phase4.functions import (
     load_datasets,
     prepare_data,
     select_variables,
