+++ famd_individuals.py
@@ -8,7 +8,7 @@ import seaborn as sns
 import pandas as pd
 import numpy as np
 
-import phase4_functions as pf
+import phase4.functions as pf
 
 
 def load_dataset(path: Path) -> pd.DataFrame:
