+++ generate_distance_matrix.py
@@ -10,7 +10,7 @@ import yaml
 from scipy.spatial.distance import pdist, squareform
 
 import phase4
-from phase4_functions import (
+from phase4.functions import (
     load_datasets,
     prepare_data,
     select_variables,
