+++ tests/test_compare_datasets_versions.py
@@ -1,5 +1,5 @@
 import pandas as pd
-import phase4_functions as pf
+import phase4.functions as pf
 
 
 def test_compare_datasets_versions_outlier(tmp_path):
