+++ tests/test_unsupervised_cv.py
@@ -1,6 +1,6 @@
 import pandas as pd
 import numpy as np
-import phase4_functions as pf
+import phase4.functions as pf
 
 
 def generate_data(shift: float = 0.0, n: int = 200, seed: int = 0):
